modal recipes
