+++ src/sass/components/_popular.scss
@@ -1,3 +1,7 @@
+
+
+
+
 .section-container {
   position: relative;
   margin: 0 auto;
@@ -47,7 +51,6 @@
 
 .popular-recipes-item {
   display: flex;
-  // justify-content: space-between;
   gap: 16;
   margin-bottom: 24px;
   p {
@@ -65,37 +68,6 @@
 .popular-recipes-img {
   border-radius: 5px;
   max-width: none;
-}
-
-.loader {
-  width: 120px;
-  height: 80px;
-  position: absolute;
-  top: 90px;
-  transform: rotate(-90deg);
-  background: linear-gradient(174deg, #0000 49%, #000 50%, #0000 51%);
-}
-.loader:after,
-.loader:before {
-  content: '';
-  border-radius: 100%;
-  width: 35px;
-  height: 35px;
-  display: block;
-  position: absolute;
-  border: 4px dashed #fff;
-  bottom: 49px;
-  transform: rotate(0deg);
-  box-sizing: border-box;
-  animation: tape 4s linear infinite;
-}
-.loader:before {
-  right: -14px;
-  box-shadow: 0 0 0 4px #fff, 0 0 0 34px #000, 0 0 5px 34px #0005;
-}
-.loader:after {
-  left: -13px;
-  box-shadow: 0 0 0 4px #fff, 0 0 0 65px #000, 0 0 5px 65px #0005;
 }
 
 @keyframes tape {
@@ -159,10 +131,6 @@
   .section-container {
     width: 100%;
   }
-
-  // .popular-recipes-item{
-  //   justify-content: center;
-  // }
 
   .popular-recipes-item {
     margin-bottom: 32px;
